Test delete category service

+++ Back-end/services/delete_category.py
@@ -3,4 +3,5 @@ from data.categories import Category
 def delete_category(id):
     category = Category.get_by_id(id)
     
-    category.delete_instance()
+    if category is not None:
+        category.delete_instance()
